requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-role-permissions==3.2.0
idna==3.7
phonenumbers==8.13.36
pillow==10.3.0
pytz==2024.1
requests==2.31.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
validate-docbr==1.10.0
